services/serve: reorganize structure and rename to chain
