+++ starport/interface/cli/starport/cmd/build.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/tendermint/starport/starport/pkg/gomodulepath"
-	starportserve "github.com/tendermint/starport/starport/services/serve"
+	"github.com/tendermint/starport/starport/services/chain"
 )
 
 func NewBuild() *cobra.Command {
@@ -29,7 +29,7 @@ func buildHandler(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	app := starportserve.App{
+	app := chain.App{
 		Name: path.Root,
 		Path: appPath,
 	}
@@ -44,7 +44,7 @@ func buildHandler(cmd *cobra.Command, args []string) error {
 		cancel()
 	}()
 
-	s, err := starportserve.New(app, verbose)
+	s, err := chain.New(app, verbose)
 	if err != nil {
 		return err
 	}
