+++ starport/interface/cli/starport/cmd/relayer.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/tendermint/starport/starport/pkg/gomodulepath"
-	starportserve "github.com/tendermint/starport/starport/services/serve"
+	"github.com/tendermint/starport/starport/services/chain"
 )
 
 func NewRelayer() *cobra.Command {
@@ -42,12 +42,12 @@ func relayerInfoHandler(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	app := starportserve.App{
+	app := chain.App{
 		Name: path.Root,
 		Path: appPath,
 	}
 
-	s, err := starportserve.New(app, false)
+	s, err := chain.New(app, false)
 	if err != nil {
 		return err
 	}
@@ -64,12 +64,12 @@ func relayerAddHandler(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return err
 	}
-	app := starportserve.App{
+	app := chain.App{
 		Name: path.Root,
 		Path: appPath,
 	}
 
-	s, err := starportserve.New(app, false)
+	s, err := chain.New(app, false)
 	if err != nil {
 		return err
 	}
