+++ starport/services/chain/app.go
@@ -1,4 +1,4 @@
-package starportserve
+package chain
 
 import (
 	"os"
