+++ starport/services/chain/dev.go
@@ -1,4 +1,4 @@
-package starportserve
+package chain
 
 import (
 	"context"
