+++ starport/services/chain/plugin-launchpad.go
@@ -1,4 +1,4 @@
-package starportserve
+package chain
 
 import (
 	"context"
@@ -12,7 +12,7 @@ import (
 	"github.com/tendermint/starport/starport/pkg/cmdrunner/step"
 	"github.com/tendermint/starport/starport/pkg/cosmosver"
 	"github.com/tendermint/starport/starport/pkg/xurl"
-	starportconf "github.com/tendermint/starport/starport/services/serve/conf"
+	starportconf "github.com/tendermint/starport/starport/services/chain/conf"
 )
 
 type launchpadPlugin struct {
